Move detected result visualization to a separate file
